refactor: 주사위 컴포넌트 및 이동로직 GameBoard에서 분리
